simplified the sample to match the forge.learning tutorial.
