--- config.json
@@ -1,5 +1,0 @@
-{
-  "FORGE_CLIENT_ID": "",
-  "FORGE_CLIENT_SECRET": "",
-  "PORT": "3000"
-}
